+++ framework/validator/src/zod/types/schema.types.ts
@@ -92,12 +92,12 @@ export type UnionZodResolve<T extends ZodUnionContainer> = T extends [
   : [ZodNever, ZodNever];
 
 /**
- * Resolves a Zod schema type T to its resolved type. The depth is limited to 31 to prevent infinite recursion.
+ * Resolves a Zod schema type T to its resolved type. The depth is limited to 29 to prevent infinite recursion.
  *
  * @template T - The Zod schema type to resolve.
  * @template Depth - The current depth of the resolution.
  */
-export type ZodResolve<T, Depth extends number = 0> = Depth extends 30
+export type ZodResolve<T, Depth extends number = 0> = Depth extends 29
   ? ZodUnknown
   : T extends LiteralSchema
     ? ZodLiteral<T>
@@ -106,5 +106,7 @@ export type ZodResolve<T, Depth extends number = 0> = Depth extends 30
       : T extends UnboxedZodObjectSchema
         ? ZodObject<{
             [K in keyof T]: ZodResolve<T[K], Increment<Depth>>;
-          }>
+          }> extends infer R
+          ? R
+          : ZodNever
         : ZodNever;