+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/hyper-express
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/validator@0.4.12
+  - @forklaunch/core@0.6.4
+
 ## 0.4.3
 
 ### Patch Changes